+++ jsf/src/main/java/org/jboss/as/jsf/deployment/JandexAnnotationProvider.java
@@ -20,13 +20,16 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.web.deployment.jsf;
+package org.jboss.as.jsf.deployment;
 
-import com.sun.faces.spi.AnnotationProvider;
 import java.lang.annotation.Annotation;
 import java.util.Map;
 import java.util.Set;
+
 import javax.servlet.ServletContext;
+
+import com.sun.faces.spi.AnnotationProvider;
+import org.jboss.as.web.deployment.jsf.JsfAnnotationProcessor;
 
 /**
  * {@link }AnnotationProvider} implementation which provides the JSF annotations which we parsed from from a
@@ -35,14 +38,13 @@ import javax.servlet.ServletContext;
  * @author John Bailey
  */
 public class JandexAnnotationProvider extends AnnotationProvider {
-    static final String FACES_ANNOTATIONS = "FACES_ANNOTATIONS";
 
     private final Map<Class<? extends Annotation>, Set<Class<?>>> annotations;
 
     @SuppressWarnings("unchecked")
     public JandexAnnotationProvider(final ServletContext servletContext) {
         super(servletContext);
-        annotations = (Map<Class<? extends Annotation>, Set<Class<?>>>) servletContext.getAttribute(FACES_ANNOTATIONS);
+        annotations = (Map<Class<? extends Annotation>, Set<Class<?>>>) servletContext.getAttribute(JsfAnnotationProcessor.FACES_ANNOTATIONS);
     }
 
     // Note: The Mojarra 2.0 SPI specifies that this method takes Set<URL> as its argument.  The Mojarra 2.1 SPI
