+++ jsf/src/main/resources/META-INF/services/com.sun.faces.spi.annotationprovider
@@ -1,0 +1,1 @@
+org.jboss.as.jsf.deployment.JandexAnnotationProvider
