+++ jsf/src/main/resources/META-INF/services/com.sun.faces.spi.injectionprovider
@@ -1,0 +1,1 @@
+org.jboss.as.jsf.deployment.JsfInjectionProvider:org.jboss.as.jsf.deployment.JsfInjectionProvider
