--- web/src/main/resources/META-INF/services/com.sun.faces.spi.annotationprovider
@@ -1,1 +1,0 @@
-org.jboss.as.web.deployment.jsf.JandexAnnotationProvider
